Created a Code folder for keeping all the C# code files. Moved those files there. Also added third tab for a third chart example which i will work on next.
