+++ Chart/MainWindow.xaml.cs
@@ -1,4 +1,5 @@
-﻿using System;
+﻿using Chart.Code;
+using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
